Here's the refactored commit message:
refactor: Modularize MCP server tools

This commit refactors the `mcp/calculater_mcp.py` file by moving tool
definitions into categorized modules within a new `mcp/tools/` directory.

Key changes:
- Created `mcp/tools/` directory with `__init__.py`.
- Created `mcp/__init__.py` to define `mcp` as a package.
- Moved tool definitions into new modules:
    - `mcp/tools/arithmetic.py` (add, subtract, multiply, divide)
    - `mcp/tools/advanced_math.py` (power, sqrt, get_constant)
    - `mcp/tools/trigonometry.py` (sin, cos, tan, asin_op, acos_op, atan_op)
    - `mcp/tools/calculus.py` (differentiate, integrate_indefinite)
    - `mcp/tools/plotting.py` (plot_expression)
- Each new tool module now handles its own specific imports (math, sympy,
  matplotlib, etc.) and logging setup (`logger = logging.getLogger(__name__)`).
- Tools in these modules import the `mcp` instance from
  `calculater_mcp.py` to use the `@mcp.tool()` decorator.
- `mcp/calculater_mcp.py` is now significantly leaner. It defines the
  `FastMCP` instance and imports all tool functions from the new
  modules to register them. Redundant imports have been removed.
- Updated the "Project Structure" section in `README.md` to reflect
  these changes.

This refactoring improves code organization, readability, and
maintainability of the MCP server tools.
