+++ mcp/__init__.py
@@ -1,0 +1,1 @@
+# This file makes the 'mcp' directory a Python package.
