+++ mcp/tools/__init__.py
@@ -1,0 +1,1 @@
+# This file makes the 'tools' directory a Python package.
